Crate moonbeam_primitives_ext

Environmental-aware externalities for EVM tracing in Wasm runtime. This enables capturing the - potentially large - trace output data in the host and keep a low memory footprint in --execution=wasm.

The original trace Runtime Api call is wrapped using environmental (thread local).

Arguments are scale-encoded known types in the host.

Host functions will decode the input and emit an event with environmental.

Modules

moonbeam_ext