moonbeam_primitives_ext

Module moonbeam_ext

Structs

HostFunctions

Provides implementations for the extern host functions.

Functions

call_list_entry

call_list_new

An event to create a new CallList (currently a new transaction when tracing a block).

evm_event

An Evm event proxied by the Moonbeam runtime to this host function. evm -> moonbeam_runtime -> host.

gasometer_event

A Gasometer event proxied by the Moonbeam runtime to this host function. evm_gasometer -> moonbeam_runtime -> host.

raw_gas

raw_return_value

raw_step

runtime_event

A Runtime event proxied by the Moonbeam runtime to this host function. evm_runtime -> moonbeam_runtime -> host.

step_event_filter

Allow the tracing module in the runtime to know how to filter Step event content, as cloning the entire data is expensive and most of the time not necessary.